+++ packages/core/button/src/components/__tests__/unit/server-side-hydrate.tsx
@@ -34,7 +34,6 @@ test('should ssr then hydrate button correctly', async () => {
           s === 'style'
         ),
     );
-
-    expect(mockCalls.length).toBe(0); // eslint-disable-line no-console
+    expect(mockCalls.length).toBe(0);
   });
 });