+++ packages/media/smart-card/src/state/store/__tests__/index.test.tsx
@@ -1,8 +1,9 @@
 import { renderSmartLinkHook } from '../../../utils/test-utils';
 import { useSmartCardState } from '..';
 import { CardStore } from '../../types';
-
-describe('useSmartCardState()', () => {
+// TODO: Those tests are wrong and are not entering into the renderHook callbacks.
+// https://product-fabric.atlassian.net/browse/SL-366
+describe.skip('useSmartCardState()', () => {
   let mockUrl = 'some.url';
 
   it('correctly returns default state', () => {