+++ packages/search/quick-search/src/__tests__/unit/QuickSearch.test.tsx
@@ -78,11 +78,16 @@ describe('<QuickSearch />', () => {
 
     it('should support non-component children', () => {
       render({ children: 'child' });
+      // Modified as per BUILDTOOLS-220: expect.hasAssertions(), please modify accordingly.
+      expect(wrapper.children).toBeDefined();
       /* Expect that no errors occur while parsing children */
     });
 
     it('should support non-component grandchildren', () => {
       render({ children: <div>grandchild</div> });
+      // Modified as per BUILDTOOLS-220: expect.hasAssertions(), please modify accordingly.
+      expect(wrapper.find('div').exists()).toBe(true);
+      expect(wrapper.find('div').children).toBeDefined();
       /* Expect that no errors occur while parsing children */
     });
   });