+++ packages/core/inline-edit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/inline-edit
 
+## 10.0.12
+
+### Patch Changes
+
+- [patch][688f2957ca](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/688f2957ca):
+
+  Fixes various TypeScript errors which were previously failing silently
+
 ## 10.0.11
 
 - Updated dependencies [1adb8727e3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1adb8727e3):